+++ frontend/src/stores/slices/requestSlice.ts
@@ -21,66 +21,55 @@ export const createRequestSlice: StateCreator<
   setMethod: (method) =>
     set((state) => {
       state.method = method;
-      state.isDirty = true;
     }),
 
   setUrl: (url) =>
     set((state) => {
       state.url = url;
-      state.isDirty = true;
     }),
 
   setRequestHeaders: (headers) =>
     set((state) => {
       state.requestHeaders = headers;
-      state.isDirty = true;
     }),
 
   setRequestBody: (body) =>
     set((state) => {
       state.requestBody = body;
-      state.isDirty = true;
     }),
 
   setResponseHeaders: (headers) =>
     set((state) => {
       state.responseHeaders = headers;
-      state.isDirty = true;
     }),
 
   setResponseBody: (body) =>
     set((state) => {
       state.responseBody = body;
-      state.isDirty = true;
     }),
 
   updateRequestHeader: (key, value) =>
     set((state) => {
       state.requestHeaders[key] = value;
-      state.isDirty = true;
     }),
 
   removeRequestHeader: (key) =>
     set((state) => {
       delete state.requestHeaders[key];
-      state.isDirty = true;
     }),
 
   updateResponseHeader: (key, value) =>
     set((state) => {
       state.responseHeaders[key] = value;
-      state.isDirty = true;
     }),
 
   removeResponseHeader: (key) =>
     set((state) => {
       delete state.responseHeaders[key];
-      state.isDirty = true;
     }),
 
   resetRequest: () =>
     set((state) => {
       Object.assign(state, DEFAULT_REQUEST_STATE);
-      state.isDirty = true;
     }),
 });